Add files via upload
I wrote a combinatorial analysis program. I had the ability. I am grateful. I named the combinatorial analysis program. The name of the combinatorial analysis program is BIGS907. I posted the code for BIGS907 elsewhere in this repository. BIGS907 enables me. I input a predetermined set of game rules for a comparing card game belonging to the Finnish 27 family wherein; both cards in the dealer's initial hand are exposed. Accordingly, the player makes a game wager. The dealer deals a set of initial hands. The set of initial hands consists of a player's hand and a dealer's hand. Both hands consist of two cards. All cards are dealt face up.  BIGS907 finds the basic strategy. The basic strategy loses the least amount of game wager to the house in the long run with strictly average luck. BIGS907 calculates an estimate of the expected value of the game. For the purposes of calculating the expected value of the game, BIGS907 assumes. The player always uses basic strategy when making decisions on how to play the player's hand.  The house edge of a game is equal to negative one times the expected value of the game.

I used the BIGS907 software application. I found the basic strategy and expected value of a comparing card game belonging to the Finnish 27 family of table card games. I specified a predetermined set of game rules for the comparing card game. The predetermined set of games specifies. The playing cards have numerical values. A hand total is equal to the sum of the numerical values in a hand. An Ace may be either high or low. A high ace has a numerical value of 14. A low ace has a numerical value of 1. Let us suppose. A hand includes a high ace. In that event, the hand is called a soft hand. Let us suppose. A hand does not include a high ace. In that event, the hand is called a hard hand. A soft hand has a soft hand total. The soft hand total is preceded by a prefix. The prefix is "soft". For example, "soft 24" is a soft hand total. A hard hand has a hard hand total. The hard hand total is preceded by a prefix. The prefix is "hard". For example, Hard 21 is a hard hand total. The dealer plays the dealer's hand in accordance with a predetermined set of game rules. The predetermined set of game rules are. Let us suppose. The dealer's hand total is less than hard 21. In that event, the dealer must add an additional card to the dealer's hand. Let us suppose. The dealer's hand total is at least equal to hard 21. In that event, the dealer must stand. The dealer's hand is complete. Let us suppose. The dealer's hand total is less than soft 24. In that event, the dealer must add one additional card to the dealer's hand. Let us suppose. The dealer's hand total is at least equal soft 24. In that event, the dealer must stand. The dealer's hand is complete. Accordingly, the dealer must stand with hard 21 or better. The dealer must stand with soft 24 or better. The dealer uses one deck of playing cards. The one deck of playings cards includes two jokers. For that reason, I named this folder J21h24s1d. 

game7i.text contains a record user-specified input. game7o.txt contains a set of combination dependent basic strategy tables. The set of combination dependent basic strategy tables displays the best strategy for the play of the player's hand given each of the possible combinations of cards in a set of hands consisting of the player's initial hand and the dealer's initial hand.  game7o.txt contains a set of combination dependent expected value tables. The set of combination dependent expected value tables displays the expect value of the best strategy for the play of the player's hand given each of the possible combinations. 

I wrote another software application. I had the ability. I am grateful. I named the software application BIGS921. I posted the code for BIGS921 elsewhere in this repository. BIGS921 takes input from BIGS907. The input specifies data about a set of combination dependent basic strategy tables. The player makes a game wager. The dealer deals a set of initial hands. The initial hands consists of a player's hand and a dealer's hand. Both initial hands consist of two cards. All playing cards are dealt face up. The set of combination dependent basic strategy tables specify the basic strategy for the play of the player's hand given the predetermined set of game rules and given the numerical values of the playing cards in the set of initial hands.  BIGS921 uses the data. BIGS921 converts the set of combinations dependent basic strategy tables into a total dependent basic strategy tables. The player's initial hand has a hand total. The sum of the numerical values of the two playing cards in the player's initial hand is the hand total. The dealer's initial hand has a hand total. The sum of the numerical values of the two playing cards in the dealer's initial hand is the hand total. The total dependent basic strategy table specifies the basic strategy for the play of the player's hand given the user-defined set of game rules and given all possible the hand totals. The number of possible hand totals is far less than the number of possible combinations of card values. For that reason, total-dependent basic strategy table is far more compact and concise than the set of combination dependent basic strategy tables. 

I used BIGS907. I specified the predetermined set of game rules for game 7. BIGS907 output a set of combination dependent basic strategy tables, a set of combination dependent expected value tables. BIGS907 output the tables as text files. BIGS907 output the tables as data files. I used BIGS921. I loaded the data for game7 into program memory. BIGS921 used the data. BIGS921 converted the combination dependent basic strategy tables into total dependent basic strategy tables. 

game7TDAi.text contains a record of user-input. game7TDAd.text displays the total-dependent basic strategy tables in the form of a set of combination dependent basic strategy tables. BIGS921 makes changes to the tables output by BIGS907.  The combination dependent basic strategy tables enables me. I compare tables output by BIGS907 to tables output by BIGS921. I easily identify changes. These changes are exception rules. The exception rules apply to the total dependent basic strategy table. game7TDAh.html displays the total dependent basic strategy table. 

Let us suppose. You have read this far. In that event, you have been generous with your time and energy. Thank you for your generosity. Your generosity is an inspiration. You inspire me. I wish us happiness and good health. I offer you my best regards.